+++ packages/@headlessng-examples/src/lib/molecules/checkbox-field/checkbox-field.component.ts
@@ -9,9 +9,17 @@ import {
   FieldComponent,
   LabelComponent
 } from '../../atoms';
+import { ErrorMessagesComponent } from '../error-messages';
 
 @Component({
-  imports: [CommonModule, CheckboxComponent, DescriptionComponent, FieldComponent, LabelComponent],
+  imports: [
+    CommonModule,
+    CheckboxComponent,
+    DescriptionComponent,
+    ErrorMessagesComponent,
+    FieldComponent,
+    LabelComponent
+  ],
   selector: 'h-checkbox-field',
   templateUrl: 'checkbox-field.component.html',
   styleUrl: 'checkbox-field.component.css',
@@ -21,4 +29,5 @@ export class CheckboxFieldComponent {
   public readonly control = input.required<FormControl<CheckboxChecked>>();
   public readonly description = input<string | undefined>();
   public readonly label = input.required<string>();
+  public readonly required = input<boolean>(false);
 }