+++ packages/@headlessng-primitives/description/src/lib/description.directive.ts
@@ -1,12 +1,9 @@
 import { Directive } from '@angular/core';
-import { DescriptionFieldRef } from '@headlessng/primitives/field';
+import { DescriptionElement } from '@headlessng/primitives/field';
 
 @Directive({
   exportAs: 'hDescriptionRef',
-  host: {
-    '[attr.id]': 'id()'
-  },
   selector: '[hDescription]',
   standalone: true
 })
-export class DescriptionDirective extends DescriptionFieldRef {}
+export class DescriptionDirective extends DescriptionElement {}