+++ packages/@headlessng-primitives/disabled/src/lib/disabled.directive.ts
@@ -1,13 +1,4 @@
-import {
-  booleanAttribute,
-  Directive,
-  effect,
-  inject,
-  Injector,
-  input,
-  output,
-  signal
-} from '@angular/core';
+import { Directive, model } from '@angular/core';
 
 @Directive({
   exportAs: 'hDisabledRef',
@@ -20,36 +11,9 @@ import {
   standalone: true
 })
 export class DisabledDirective {
-  private readonly _injector = inject(Injector);
-
-  private readonly _disabled = signal<boolean>(false);
-  private readonly _disabledEffect = effect(
-    () => {
-      this.disabledChange.emit(this._disabled());
-    },
-    {
-      injector: this._injector
-    }
-  );
-
-  public readonly disabled = this._disabled.asReadonly();
-  public readonly disabledChange = output<boolean>();
-
-  public readonly _disabledInput = input(this._disabled(), {
-    alias: 'disabled',
-    transform: booleanAttribute
-  });
-  private readonly _disabledInputEffect = effect(
-    () => {
-      this._disabled.set(this._disabledInput());
-    },
-    {
-      allowSignalWrites: true,
-      injector: this._injector
-    }
-  );
+  public readonly disabled = model(false);
 
   public setDisabled(disabled: boolean): void {
-    this._disabled.set(disabled);
+    this.disabled.set(disabled);
   }
 }