+++ packages/@headlessng-primitives/error-message/src/lib/error-message.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from '@angular/core';
+import { ErrorMessageElement } from '@headlessng/primitives/field';
+
+@Directive({
+  exportAs: 'hErrorMessageRef',
+  selector: '[hErrorMessage]',
+  standalone: true
+})
+export class ErrorMessageDirective extends ErrorMessageElement {}