+++ packages/@headlessng-primitives/label/src/lib/label.directive.ts
@@ -1,12 +1,9 @@
 import { Directive } from '@angular/core';
-import { LabelFieldRef } from '@headlessng/primitives/field';
+import { LabelElement } from '@headlessng/primitives/field';
 
 @Directive({
   exportAs: 'hLabelRef',
-  host: {
-    '[attr.id]': 'id()'
-  },
   selector: 'label[hLabel]',
   standalone: true
 })
-export class LabelDirective extends LabelFieldRef {}
+export class LabelDirective extends LabelElement {}